pyproject.toml
[tool.poetry]
name = "pooper-detection"
version = "0.1.0"
description = ""
authors = ["DantasB <bruno.d.paiva@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.9.0"
pillow = "^10.3.0"
opencv-python = "^4.10.0.84"
pyyaml = "^6.0.1"
scipy = "^1.13.1"
requests = "^2.32.3"
torch = "^2.3.1"
torchvision = "^0.18.1"
tqdm = "^4.66.4"
numpy = "1.26.3"
pygame = "^2.5.2"
ffmpeg-python = "^0.2.0"
psutil = "^5.9.8"
pandas = "^2.2.2"
ultralytics = "^8.2.35"
pytelegrambotapi = "^4.19.2"
black = "^24.4.2"
ruff = "^0.4.9"
mypy = "^1.10.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--disable-warnings --cov=app/src --cov=app/tests --cov-report=term-missing -v -rsxX"

[tool.coverage.run]
branch = true
source = ["app/src", "app/tests"]

[tool.coverage.report]
show_missing = true
